Put 2fa for tests into helper fx for cleaner code
