+++ server/test/integration/api/ambulances.js
@@ -18,18 +18,9 @@ describe('/api/ambulances', () => {
       .post('/auth/local/login')
       .set('Accept', 'application/json')
       .send({ username: 'sutter.operational@example.com', password: 'abcd1234' });
-
-    // Call the two-factor authentication endpoint
-    await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-    const sentMail = nodemailermock.mock.sentMail();
-    // Extract authentication code from the sent email
-    const regex = /Authentication Code: (\d{6})/;
-    const match = regex.exec(sentMail[0].text);
-    const authCode = match[1];
-    // Submit the authentication code
-    await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+    await helper.twoFactorAuthSession(testSession);
   });
-  after(async () => {
+  afterEach(async () => {
     nodemailermock.mock.reset();
   });
