+++ server/test/integration/api/ringdowns.js
@@ -40,16 +40,7 @@ describe('/api/ringdowns', () => {
         .set('Accept', 'application/json')
         .send({ username: 'super.user@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
-
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+      await helper.twoFactorAuthSession(testSession);
 
       const response = await testSession.get('/api/ringdowns').set('Accept', 'application/json').expect(HttpStatus.OK);
       assert.deepStrictEqual(response.body.length, 5);
@@ -62,16 +53,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'sutter.operational@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession
         .get('/api/ringdowns')
         .query({ hospitalId: '7f666fe4-dbdd-4c7f-ab44-d9157379a680' })
@@ -88,16 +70,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'amr.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession.get('/api/ringdowns/mine').set('Accept', 'application/json').expect(HttpStatus.OK);
       assert.deepStrictEqual(response.body.length, 1);
       assert.deepStrictEqual(response.body[0].id, '4889b0c8-ce48-474a-ac5b-c5aca708451c');
@@ -138,16 +111,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'norcal.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession
         .post('/api/ringdowns')
         .set('Accept', 'application/json')
@@ -182,17 +146,7 @@ describe('/api/ringdowns', () => {
         .set('Accept', 'application/json')
         .send({ username: 'norcal.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
-
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession
         .post('/api/ringdowns')
         .set('Accept', 'application/json')
@@ -256,17 +210,7 @@ describe('/api/ringdowns', () => {
         .set('Accept', 'application/json')
         .send({ username: 'norcal.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
-
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession
         .post('/api/ringdowns')
         .set('Accept', 'application/json')
@@ -302,16 +246,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'norcal.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       const response = await testSession
         .post('/api/ringdowns')
         .set('Accept', 'application/json')
@@ -350,16 +285,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'sutter.operational@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
-
+      await helper.twoFactorAuthSession(testSession);
       let now = new Date();
       await testSession
         .patch('/api/ringdowns/d4fd2478-ecd6-4571-9fb3-842bfc64b511/deliveryStatus')
@@ -467,15 +393,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'king.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+      await helper.twoFactorAuthSession(testSession);
 
       const now = new Date();
       await testSession
@@ -530,15 +448,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'sffd.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+      await helper.twoFactorAuthSession(testSession);
 
       const response = await testSession
         .patch('/api/ringdowns/8b95ea8a-0171-483a-be74-ec17bbc12247')
@@ -588,15 +498,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'sffd.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+      await helper.twoFactorAuthSession(testSession);
 
       const response = await testSession
         .get('/api/ringdowns/8b95ea8a-0171-483a-be74-ec17bbc12247')
@@ -658,15 +560,7 @@ describe('/api/ringdowns', () => {
         .send({ username: 'second.sffd.paramedic@example.com', password: 'abcd1234' })
         .expect(HttpStatus.OK);
 
-      // Call the two-factor authentication endpoint
-      await testSession.get('/auth/local/twoFactor').set('Accept', 'application/json');
-      const sentMail = nodemailermock.mock.sentMail();
-      // Extract authentication code from the sent email
-      const regex = /Authentication Code: (\d{6})/;
-      const match = regex.exec(sentMail[0].text);
-      const authCode = match[1];
-      // Submit the authentication code
-      await testSession.post('/auth/local/twoFactor').set('Accept', 'application/json').send({ code: authCode });
+      await helper.twoFactorAuthSession(testSession);
 
       await testSession.get('/api/ringdowns/8b95ea8a-0171-483a-be74-ec17bbc12247').set('Accept', 'application/json').expect(HttpStatus.OK);
     });
